Finally, some functional (5V only, unfortunately) logic gates!!
5V works, but 2.5V would sure be nicer. Lower power, etc. Maybe there's a compromise somewhere in the middle. At any rate, these gates are ~100mA per gate, but have 10ns propagation delay, so maybe there's room for savings.

The temperature dependency is not great, but acceptable. What's interesting is that the *first* inverter in the chain develops some temperature sensitivity in it's trigger point, but it doesn't seem to get amplified downwards an inverter chain. In other words, it's a self-limiting phenomena, and might not even matter when every driver is of the same kind as opposed to an ideal voltage source in LTSpice.