Finally figured out temperature compensation.
This circuit now works well at 5V, though current consumption is high.

I also need now 4 (!) power supplies, fine-tuned to compensate for Vbe drop.
This is for now represented by ideal voltage supplies, but of course it
will need to be something that needs to be implemented by
real circuits in the end.

But, at least it allows me to make forward progress.

NOTE: 2.5V operation is not possible, the emitter follower drops
are too high.