I don't think I can get a better bit-cell then this.
Here we have a 5V, ~20ns, relatively low power bit-cell, where the two inverters and the shift-enable logic total needs only 6 transistors and 2 diodes. I'm not sure I would want to use the same design for generic logic, but I think it's going to be great for the shift-registers.

For things with parallel load requirements yet another structure (with 2 more transistors to make the input OR gates less hacky) might be needed. Or maybe not. We shall see.