All tested instructions execute properly now
Unfortunately I needed to add another mux path: L_BUS_D needs
to load from both ALU_RESULT (for SWAP) and from L_ALU_RESULT
(for everything else). This is because L_BUS_D is also an input to most
ALU operations, so the result has to go through L_ALU_RESULT for a
cycle, but SWAP can't cycle it's result through there as L_ALU_RESULT
is a temporary storage for the register content.

Also MOV-s are somewhat special in selecting their ALU inputs as one or
the other needs to be zero-ed out based on what the move direction is.
BTW: MOV could easily be a UNARY op, freeing up a BINARY encoding slot. This is not done though.