Full 16-bit ALU out of transistors.
The propagation delay seems to be pretty
good, actually:

~220ns at 125C and (thought the sim is still running)
~350ns at -25C

This is well within target for 400ns and
kind of surprising for a ripple-carry
implementation. However if it holds,
this makes things much easier on the
implementation.

BTW: ROL/ROR bit-forwarding is hooked
up incorrectly on the top, so that needs
to be fixed.

Also, only ADD is tested at the moment.
The rest should be easy as mostly can be
tested on the bit level.