A workable sense-amp design.
This can probably be fine-tuned a little to
get for instance power consumption
under control, but otherwise, pretty
good.

The design is DC-coupled.

It has about 85ns delay to it, but I don't
think that matters all that much.

It's threshold changes about 1mV
with temperature.

Rise times are ~17ns, fall times are ~25ns.

All are very acceptable values.

Complexity wise it has 16 transistors,
3 diodes and ~25 resistors.