Finally (in version 3) a decent current source.
This current source  seems to be well
enough tuned to the load switches that
that I have. It is parameterized to the inhibit
currents at the moment, I'll have to tune
a version for the line drivers as well.

Rise times are workable with no inductive
loads, have to test what it would look like
with worst-case loads.