Much more updated current source.
This one is back to four outputs and an
extra buffer stage to make the crosstalk
nonexistence.

The penalty for it is a few extra
transistors and more importantly a
diminished temperature independence.
Still not awful and after thermistor
compensation, pretty good, actually.

Also, the load regulation could be better,
something to look into with the actual
load.