I managed to optimize out 4 transistors per drive path.
Also fixed shoot-through. So now, each X-Y driver needs
16 transistors, needing only 768 total. (BTW, I think
previously I estimated double the transistor count
than actually needed, so it went from 1100 to 768).

I also gotten rid of the shoot-through that was due
to the not-so-clean high level on the high voltage
drive of the high-side.

Now the price I pay is rather sluggish rise-time, we
barely reach the target current by the end of the 200ns
on-time.