Spent a lot of time with common-base input stages.
Not sure I got to where I want to be, but there is *something*. About 150uV output swing with
a significantly larger temperature-dependent offset (even though the circuit is now pretty well
temperature compensated). It's also rather complicated at 5 transistors. I'm wondering though
if the emitter followers are needed, or if the common-base input can drive the diff-amp directly...